+++ gotu/quasar.py
@@ -1,5 +1,4 @@
-"""
-=======================================================================
+"""=======================================================================
  What would light emerging at the event horizon of a quasar look like?
 =======================================================================
 
@@ -18,5 +17,13 @@ In a place where the passage of time is close to zero, but not actually, zero.
 
 JWST
 ====
+
+The model for quasars outlined in `gotu`_ requires just two parameters.
+
+The central mass and the temperature and density of matter around the
+Eddington sphere of the quasar's black hole.
+
+There is a huge range of possible redshifts, depending on these two
+parameters.  
 
 """